支持使用身份码时前端直连
