+++ frontend/package.json
@@ -13,6 +13,7 @@
     "downloadjs": "^1.4.7",
     "element-ui": "^2.15.13",
     "lodash": "^4.17.21",
+    "pako": "^2.1.0",
     "vue": "^2.7.14",
     "vue-i18n": "^8.28.2",
     "vue-router": "^3.6.5"
