--- frontend/src/api/chat/ChatClientDirect/index.js
@@ -1,505 +1,0 @@
-import axios from 'axios'
-
-import { BrotliDecode } from './brotli_decode'
-import { getUuid4Hex } from '@/utils'
-import * as chat from '..'
-
-const HEADER_SIZE = 16
-
-// const WS_BODY_PROTOCOL_VERSION_NORMAL = 0
-// const WS_BODY_PROTOCOL_VERSION_HEARTBEAT = 1
-// const WS_BODY_PROTOCOL_VERSION_DEFLATE = 2
-const WS_BODY_PROTOCOL_VERSION_BROTLI = 3
-
-// const OP_HANDSHAKE = 0
-// const OP_HANDSHAKE_REPLY = 1
-const OP_HEARTBEAT = 2
-const OP_HEARTBEAT_REPLY = 3
-// const OP_SEND_MSG = 4
-const OP_SEND_MSG_REPLY = 5
-// const OP_DISCONNECT_REPLY = 6
-const OP_AUTH = 7
-const OP_AUTH_REPLY = 8
-// const OP_RAW = 9
-// const OP_PROTO_READY = 10
-// const OP_PROTO_FINISH = 11
-// const OP_CHANGE_ROOM = 12
-// const OP_CHANGE_ROOM_REPLY = 13
-// const OP_REGISTER = 14
-// const OP_REGISTER_REPLY = 15
-// const OP_UNREGISTER = 16
-// const OP_UNREGISTER_REPLY = 17
-// B站业务自定义OP
-// const MinBusinessOp = 1000
-// const MaxBusinessOp = 10000
-
-const AUTH_REPLY_CODE_OK = 0
-// const AUTH_REPLY_CODE_TOKEN_ERROR = -101
-
-const HEARTBEAT_INTERVAL = 10 * 1000
-const RECEIVE_TIMEOUT = HEARTBEAT_INTERVAL + (5 * 1000)
-
-let textEncoder = new TextEncoder()
-let textDecoder = new TextDecoder()
-
-export default class ChatClientDirect {
-  constructor(roomId) {
-    // 调用initRoom后初始化，如果失败，使用这里的默认值
-    this.roomId = roomId
-    this.roomOwnerUid = 0
-    this.hostServerList = [
-      { host: "broadcastlv.chat.bilibili.com", port: 2243, wss_port: 443, ws_port: 2244 }
-    ]
-
-    this.onAddText = null
-    this.onAddGift = null
-    this.onAddMember = null
-    this.onAddSuperChat = null
-    this.onDelSuperChat = null
-    this.onUpdateTranslation = null
-    this.onInteractWord = null
-
-    this.websocket = null
-    this.retryCount = 0
-    this.isDestroying = false
-    this.heartbeatTimerId = null
-    this.receiveTimeoutTimerId = null
-  }
-
-  async start() {
-    await this.initRoom()
-    this.wsConnect()
-  }
-
-  stop() {
-    this.isDestroying = true
-    if (this.websocket) {
-      this.websocket.close()
-    }
-  }
-
-  async initRoom() {
-    let res
-    try {
-      res = (await axios.get('/api/room_info', { params: {
-        roomId: this.roomId
-      } })).data
-    } catch {
-      return
-    }
-    this.roomId = res.roomId
-    this.roomOwnerUid = res.ownerUid
-    if (res.hostServerList.length !== 0) {
-      this.hostServerList = res.hostServerList
-    }
-  }
-
-  makePacket(data, operation) {
-    let body = textEncoder.encode(JSON.stringify(data))
-    let header = new ArrayBuffer(HEADER_SIZE)
-    let headerView = new DataView(header)
-    headerView.setUint32(0, HEADER_SIZE + body.byteLength)   // pack_len
-    headerView.setUint16(4, HEADER_SIZE)                     // raw_header_size
-    headerView.setUint16(6, 1)                               // ver
-    headerView.setUint32(8, operation)                       // operation
-    headerView.setUint32(12, 1)                              // seq_id
-    return new Blob([header, body])
-  }
-
-  sendAuth() {
-    let authParams = {
-      uid: 0,
-      roomid: this.roomId,
-      protover: 3,
-      platform: 'web',
-      type: 2,
-      buvid: '',
-    }
-    this.websocket.send(this.makePacket(authParams, OP_AUTH))
-  }
-
-  wsConnect() {
-    if (this.isDestroying) {
-      return
-    }
-    let hostServer = this.hostServerList[this.retryCount % this.hostServerList.length]
-    const url = `wss://${hostServer.host}:${hostServer.wss_port}/sub`
-    this.websocket = new WebSocket(url)
-    this.websocket.binaryType = 'arraybuffer'
-    this.websocket.onopen = this.onWsOpen.bind(this)
-    this.websocket.onclose = this.onWsClose.bind(this)
-    this.websocket.onmessage = this.onWsMessage.bind(this)
-  }
-
-  onWsOpen() {
-    this.sendAuth()
-    this.heartbeatTimerId = window.setInterval(this.sendHeartbeat.bind(this), HEARTBEAT_INTERVAL)
-    this.refreshReceiveTimeoutTimer()
-  }
-
-  sendHeartbeat() {
-    this.websocket.send(this.makePacket({}, OP_HEARTBEAT))
-  }
-
-  refreshReceiveTimeoutTimer() {
-    if (this.receiveTimeoutTimerId) {
-      window.clearTimeout(this.receiveTimeoutTimerId)
-    }
-    this.receiveTimeoutTimerId = window.setTimeout(this.onReceiveTimeout.bind(this), RECEIVE_TIMEOUT)
-  }
-
-  onReceiveTimeout() {
-    window.console.warn('接收消息超时')
-    this.discardWebsocket()
-  }
-
-  discardWebsocket() {
-    if (this.receiveTimeoutTimerId) {
-      window.clearTimeout(this.receiveTimeoutTimerId)
-      this.receiveTimeoutTimerId = null
-    }
-
-    // 直接丢弃阻塞的websocket，不等onclose回调了
-    this.websocket.onopen = this.websocket.onclose = this.websocket.onmessage = null
-    this.websocket.close()
-    this.onWsClose()
-  }
-
-  onWsClose() {
-    this.websocket = null
-    if (this.heartbeatTimerId) {
-      window.clearInterval(this.heartbeatTimerId)
-      this.heartbeatTimerId = null
-    }
-    if (this.receiveTimeoutTimerId) {
-      window.clearTimeout(this.receiveTimeoutTimerId)
-      this.receiveTimeoutTimerId = null
-    }
-
-    if (this.isDestroying) {
-      return
-    }
-    this.retryCount++
-    window.console.warn('掉线重连中', this.retryCount)
-    window.setTimeout(this.wsConnect.bind(this), 1000)
-  }
-
-  onWsMessage(event) {
-    if (!(event.data instanceof ArrayBuffer)) {
-      window.console.warn('未知的websocket消息类型，data=', event.data)
-      return
-    }
-    let data = new Uint8Array(event.data)
-    this.parseWsMessage(data)
-
-    // 至少成功处理1条消息
-    this.retryCount = 0
-  }
-
-  parseWsMessage(data) {
-    let offset = 0
-    let dataView = new DataView(data.buffer)
-    let packLen = dataView.getUint32(0)
-    let rawHeaderSize = dataView.getUint16(4)
-    // let ver = dataView.getUint16(6)
-    let operation = dataView.getUint32(8)
-    // let seqId = dataView.getUint32(12)
-
-    switch (operation) {
-    case OP_AUTH_REPLY:
-    case OP_SEND_MSG_REPLY: {
-      // 业务消息，可能有多个包一起发，需要分包
-      while (true) { // eslint-disable-line no-constant-condition
-        let body = new Uint8Array(data.buffer, offset + rawHeaderSize, packLen - rawHeaderSize)
-        this.parseBusinessMessage(dataView, body)
-
-        offset += packLen
-        if (offset >= data.byteLength) {
-          break
-        }
-
-        dataView = new DataView(data.buffer, offset)
-        packLen = dataView.getUint32(0)
-        rawHeaderSize = dataView.getUint16(4)
-      }
-      break
-    }
-    case OP_HEARTBEAT_REPLY: {
-      // 服务器心跳包，包含人气值，这里没用
-      this.refreshReceiveTimeoutTimer()
-      break
-    }
-    default: {
-      // 未知消息
-      let body = new Uint8Array(data.buffer, offset + rawHeaderSize, packLen - rawHeaderSize)
-      window.console.warn('未知包类型，operation=', operation, dataView, body)
-      break
-    }
-    }
-  }
-
-  parseBusinessMessage(dataView, body) {
-    let ver = dataView.getUint16(6)
-    let operation = dataView.getUint32(8)
-
-    switch (operation) {
-    case OP_SEND_MSG_REPLY: {
-
-      // 业务消息
-      if (ver == WS_BODY_PROTOCOL_VERSION_BROTLI) {
-        // 压缩过的先解压
-        body = BrotliDecode(body)
-        this.parseWsMessage(body)
-      } else {
-        // 没压缩过的直接反序列化
-        if (body.length !== 0) {
-          try {
-            body = JSON.parse(textDecoder.decode(body))
-            this.handlerCommand(body)
-          } catch (e) {
-            window.console.error('body=', body)
-            throw e
-          }
-        }
-      }
-      break
-    }
-    case OP_AUTH_REPLY: {
-      // 认证响应
-      body = JSON.parse(textDecoder.decode(body))
-      if (body.code !== AUTH_REPLY_CODE_OK) {
-        window.console.error('认证响应错误，body=', body)
-        // 这里应该重新获取token再重连的，但前端没有用到token，所以不重新init了
-        this.discardWebsocket()
-        throw new Error('认证响应错误')
-      }
-      this.sendHeartbeat()
-      break
-    }
-    default: {
-      // 未知消息
-      window.console.warn('未知包类型，operation=', operation, dataView, body)
-      break
-    }
-    }
-  }
-
-  handlerCommand(command) {
-    let cmd = command.cmd || ''
-    let pos = cmd.indexOf(':')
-    if (pos != -1) {
-      cmd = cmd.substr(0, pos)
-    }
-    // console.log(command)
-    let callback = CMD_CALLBACK_MAP[cmd]
-    if (callback) {
-      callback.call(this, command)
-    }
-  }
-
-  // TODO: 欢迎入场 ws 的信息，然后给到 Room.vue
-  async interactWordCallback(command) {
-    if (!this.onInteractWord) {
-      return
-    }
-    if (command.data.text_large) {
-      return
-    }
-    let data = command.data
-    // console.log(`interactWordCallback data 是 ${JSON.stringify(data, null, 4)}`)
-
-    data = {
-      id: getUuid4Hex(),
-      roomId: data.roomid,
-      timestamp: data.timestamp,
-      avatarUrl: await avatar.getAvatarUrl(data.uid),
-      msgType: data.msg_type,
-      authorName: data.uname,
-      medalName: data.fans_medal.medal_level === 0 ? undefined : data.fans_medal.medal_name,
-      medalLevel: data.fans_medal.medal_level === 0 ? undefined : data.fans_medal.medal_level,
-      isFanGroup: data.roomid === data.fans_medal.medal_room_id ? true : false,  // 是否是粉丝团（即粉丝勋章为当前直播间的粉丝勋章）
-      privilegeType: data.fans_medal.guard_level // 所带勋章牌子的舰队等级，0非舰队，1总督，2提督，3舰长（不一定是当前直播间的粉丝勋章）
-    }
-    this.onInteractWord(data)
-  }
-
-
-  async danmuMsgCallback(command) {
-    if (!this.onAddText) {
-      return
-    }
-    let info = command.info
-
-    let roomId, medalLevel, medalName
-    if (info[3]) {
-      roomId = info[3][3]
-      medalLevel = info[3][0]
-      medalName = info[3][1]
-    } else {
-      medalName = roomId = medalLevel = 0
-    }
-    let uid = info[2][0]
-    let isAdmin = info[2][2]
-    let privilegeType = info[7]
-    let authorType
-    if (uid === this.roomOwnerUid) {
-      authorType = 3
-    } else if (isAdmin) {
-      authorType = 2
-    } else if (privilegeType !== 0) {
-      authorType = 1
-    } else {
-      authorType = 0
-    }
-
-    let textEmoticons = this.parseTextEmoticons(info)
-
-    let data = {
-      avatarUrl: await chat.getAvatarUrl(uid),
-      timestamp: info[0][4] / 1000,
-      authorName: info[2][1],
-      authorType: authorType,
-      content: info[1],
-      privilegeType: privilegeType,
-      isGiftDanmaku: Boolean(info[0][9]),
-      authorLevel: info[4][0],
-      isNewbie: info[2][5] < 10000,
-      isMobileVerified: Boolean(info[2][6]),
-      medalName: medalName,
-      medalLevel: medalLevel,
-      isFanGroup: roomId === this.roomId ? true : false,  // 是否是粉丝团（即粉丝勋章为当前直播间的粉丝勋章）
-      id: getUuid4Hex(),
-      translation: '',
-      emoticon: info[0][13].url || null,
-      textEmoticons: textEmoticons,
-    }
-    // 增加区分表情的细节数据
-    if (info[0][13]) {
-      data.emoticonDetail = info[0][13]
-    }
-    // 存储emoji占位符和图片对应关系
-    // FIXME: 对齐原 Repo 的逻辑，注解掉 emots
-    // if (info[0][15] && info[0][15].extra) {
-    //   const extraMap = JSON.parse(info[0][15].extra)
-    //   if (extraMap.emots) {
-    //     data.emots = extraMap.emots
-    //   }
-    // }
-    
-    //   {
-    //     "[哇]": {
-    //         "emoticon_id": 211,
-    //         "emoji": "[哇]",
-    //         "descript": "[哇]",
-    //         "url": "http://i0.hdslb.com/bfs/live/650c3e22c06edcbca9756365754d38952fc019c3.png",
-    //         "width": 20,
-    //         "height": 20,
-    //         "emoticon_unique": "emoji_211"
-    //     },
-    //     "[妙]": {
-    //         "emoticon_id": 210,
-    //         "emoji": "[妙]",
-    //         "descript": "[妙]",
-    //         "url": "http://i0.hdslb.com/bfs/live/08f735d950a0fba267dda140673c9ab2edf6410d.png",
-    //         "width": 20,
-    //         "height": 20,
-    //         "emoticon_unique": "emoji_210"
-    //     }
-    //  }
-    this.onAddText(data)
-  }
-
-  parseTextEmoticons(info) {
-    try {
-      let modeInfo = info[0][15]
-      let extra = JSON.parse(modeInfo.extra)
-      if (!extra.emots) {
-        return []
-      }
-      let res = Object.values(extra.emots).map(emoticon => [emoticon.descript, emoticon.url])
-      return res
-    } catch {
-      return []
-    }
-  }
-
-  sendGiftCallback(command) {
-    if (!this.onAddGift) {
-      return
-    }
-    let data = command.data
-    //  if (data.coin_type !== 'gold') { // 白嫖不丢人
-    //    return
-    //  }
-
-    data = {
-      id: getUuid4Hex(),
-      avatarUrl: chat.processAvatarUrl(data.face),
-      timestamp: data.timestamp,
-      authorName: data.uname,
-      coinType: data.coin_type,
-      totalCoin: data.total_coin,
-      giftName: data.giftName,
-      num: data.num
-    }
-    this.onAddGift(data)
-  }
-
-  async guardBuyCallback(command) {
-    if (!this.onAddMember) {
-      return
-    }
-
-    let data = command.data
-    data = {
-      id: getUuid4Hex(),
-      avatarUrl: await chat.getAvatarUrl(data.uid),
-      timestamp: data.start_time,
-      authorName: data.username,
-      privilegeType: data.guard_level
-    }
-    this.onAddMember(data)
-  }
-
-  superChatMessageCallback(command) {
-    if (!this.onAddSuperChat) {
-      return
-    }
-
-    let data = command.data
-    data = {
-      id: data.id.toString(),
-      avatarUrl: chat.processAvatarUrl(data.user_info.face),
-      timestamp: data.start_time,
-      authorName: data.user_info.uname,
-      price: data.price,
-      content: data.message,
-      translation: ''
-    }
-    this.onAddSuperChat(data)
-  }
-
-  superChatMessageDeleteCallback(command) {
-    if (!this.onDelSuperChat) {
-      return
-    }
-
-    let ids = []
-    for (let id of command.data.ids) {
-      ids.push(id.toString())
-    }
-    this.onDelSuperChat({ ids })
-  }
-}
-
-// 设置收到 指定 cmd 对应处理的function
-const CMD_CALLBACK_MAP = {
-  DANMU_MSG: ChatClientDirect.prototype.danmuMsgCallback,
-  SEND_GIFT: ChatClientDirect.prototype.sendGiftCallback,
-  GUARD_BUY: ChatClientDirect.prototype.guardBuyCallback,
-  SUPER_CHAT_MESSAGE: ChatClientDirect.prototype.superChatMessageCallback,
-  SUPER_CHAT_MESSAGE_DELETE: ChatClientDirect.prototype.superChatMessageDeleteCallback,
-  INTERACT_WORD: ChatClientDirect.prototype.interactWordCallback,
-  // 2023-07-08 短暂出现过用 WATCHED_CHANGE 展示进房信息，后又改回 INTERACT_WORD ，现在 WATCHED_CHANGE 用来刷新 X 人看过，未开播时进房不会发送 INTERACT_WORD。 这里暂时先保留
-  WATCHED_CHANGE: ChatClientDirect.prototype.interactWordCallback
-}
