+++ frontend/src/api/chat/ChatClientRelay.js
@@ -141,7 +141,6 @@ export default class ChatClientRelay {
         id: data[11],
         translation: data[12],
         emoticon: emoticon,
-        textEmoticons: data[15],
       }
       this.onAddText(data)
       break
