+++ frontend/src/views/Room.vue
@@ -25,7 +25,8 @@ import * as pronunciation from '@/utils/pronunciation'
 import * as chatConfig from '@/api/chatConfig'
 import * as chat from '@/api/chat'
 import ChatClientTest from '@/api/chat/ChatClientTest'
-import ChatClientDirect from '@/api/chat/ChatClientDirect'
+import ChatClientDirectWeb from '@/api/chat/ChatClientDirectWeb'
+import ChatClientDirectOpenLive from '@/api/chat/ChatClientDirectOpenLive'
 import ChatClientRelay from '@/api/chat/ChatClientRelay'
 import ChatRenderer from '@/components/ChatRenderer'
 import * as constants from '@/components/ChatRenderer/constants'
@@ -275,10 +276,9 @@ export default {
         this.chatClient = new ChatClientRelay(roomKey, this.config.autoTranslate)
       } else {
         if (this.roomKeyType === 1) {
-          this.chatClient = new ChatClientDirect(this.roomKeyValue)
+          this.chatClient = new ChatClientDirectWeb(this.roomKeyValue)
         } else {
-          // TODO 支持authCode
-          // this.chatClient = new ChatClientDirect(this.roomKeyValue)
+          this.chatClient = new ChatClientDirectOpenLive(this.roomKeyValue)
         }
       }
       this.chatClient.onAddText = this.onAddText
